backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
celery==5.4.0
redis==5.0.8
fastapi-mail==1.4.1
google-generativeai==0.8.3
certifi==2024.8.30
websockets==13.1